Refactor into smaller modules while preserving CLI behavior
